--- src/strategy/template.js
@@ -1,169 +1,0 @@
-/*
- * 本模块是用于将虚拟DOM变成一个函数
- */
-var parseExpr = require('./parseExpr')
-var extractBindings = require('./extractBindings')
-var parseDelimiter = require('./parseDelimiter')
-var stringify = require('./stringify')
-var config = avalon.config
-var quote = avalon.quote
-var rident =  /^[$a-zA-Z_][$a-zA-Z0-9_]*$/
-var rstatement = /^\s*var\s+([$\w]+)\s*\=\s*\S+/
-var skips = {__local__: 1,vmode:1, dom: 1}
-
-
-function parseNodes(source, inner) {
-    //ms-important， ms-controller ， ms-for 不可复制，省得死循环
-    //ms-important --> ms-controller --> ms-for --> ms-widget --> ms-effect --> ms-if
-    var buffer = inner ? [] : ['\nvar vnodes = [];']
-
-    for (var i = 0, el; el = source[i++]; ) {
-        var vnode = parseNode(el)
-        if (el.$prepend) {
-            buffer.push(el.$prepend)
-        }
-        var append = el.$append
-        delete el.$append
-        delete el.$prepend
-        if (vnode) {
-            buffer.push(vnode + '\n')
-        }
-        if (append) {
-            buffer.push(append)
-        }
-    }
-    if (!inner) {
-        buffer.push('return vnodes\n')
-    }
-    return buffer.join('\n')
-}
-
-
-
-function parseNode(vdom) {
-    switch (vdom.nodeType) {
-        case 3:
-            if (config.rexpr.test(vdom.nodeValue)) {
-                return add(stringifyText(vdom))
-            } else {
-                return addTag(vdom)
-            }
-        case 1:
-            var copy = {
-                props: {},
-                type: vdom.type,
-                nodeType: 1
-            }
-            var bindings = extractBindings(copy, vdom.props)
-            var order = bindings.map(function (b) {
-                //将ms-*的值变成函数,并赋给copy.props[ms-*]
-                //如果涉及到修改结构,则在source添加$append,$prepend
-                avalon.directives[b.type].parse(copy, vdom, b)
-                return b.name
-            }).join(',')
-            if (order) {
-                copy.order = order
-            }
-            if (vdom.isVoidTag) {
-                copy.isVoidTag = true
-            } else {
-                if (!('children' in copy)) {
-                    var c = vdom.children
-                    if (c.length) {
-                        copy.children = '(function(){' + parseNodes(c) + '})()'
-                    } else {
-                        copy.children = '[]'
-                    }
-                }
-            }
-            if (vdom.skipContent)
-                copy.skipContent = true
-            if (vdom.skipAttrs)
-                copy.skipAttrs = true
-
-            return addTag(copy)
-        case 8:
-            var nodeValue = vdom.nodeValue
-            if (vdom.dynamic === 'for') {// 处理ms-for指令
-                if (nodeValue.indexOf('ms-for:') !== 0) {
-                    avalon.error('ms-for指令前不能有空格')
-                }
-                var copy = {
-                    dynamic: 'for',
-                    vmodel: '__vmodel__'
-                }
-                for (var i in vdom) {
-                    if (vdom.hasOwnProperty(i) && !skips[i]) {
-                        copy[i] = vdom[i]
-                    }
-                }
-
-                avalon.directives['for'].parse(copy, vdom, vdom)
-                return addTag(copy)
-            } else if (vdom.dynamic) {
-                if (nodeValue.indexOf('ms-for-end:') !== 0) {
-                    avalon.error('ms-for-end指令前不能有空格')
-                }
-                vdom.$append = addTag({
-                    nodeType: 8,
-                    type: '#comment',
-                    nodeValue: vdom.signature,
-                    key: 'traceKey'
-                }) + '\n},__local__,vnodes)\n' +
-                        addTag({
-                            nodeType: 8,
-                            type: "#comment",
-                            signature: vdom.signature,
-                            nodeValue: "ms-for-end:"
-                        }) + '\n'
-                return ''
-
-            } else if (nodeValue.indexOf('ms-js:') === 0) {//插入JS声明语句
-                var statement = parseExpr(nodeValue.replace('ms-js:', ''), 'js') + '\n'
-                var ret = addTag(vdom)
-                var match = statement.match(rstatement)
-                if (match && match[1]) {
-                    vdom.$append = (vdom.$append || '') + statement +
-                            "\n__local__." + match[1] + ' = ' + match[1] + '\n'
-                } else {
-                    avalon.warn(nodeValue + ' parse fail!')
-                }
-                return ret
-            } else {
-                return addTag(vdom)
-            }
-        default:
-            if (Array.isArray(vdom)) {
-                vdom.$append = parseNodes(vdom, true)
-            }
-
-    }
-
-}
-
-module.exports = parseNodes
-
-function wrapDelimiter(expr) {
-    return rident.test(expr) ? expr : parseExpr(expr, 'text')
-}
-
-function add(a) {
-    return 'vnodes.push(' + a + ');'
-}
-function addTag(obj) {
-    return add(stringify(obj))
-}
-
-function stringifyText(el) {
-    var array = parseDelimiter(el.nodeValue)//返回一个数组
-    var nodeValue = ''
-    if (array.length === 1) {
-        nodeValue = wrapDelimiter(array[0].expr)
-    } else {
-        var token = array.map(function (el) {
-            return el.type ? wrapDelimiter(el.expr) : quote(el.expr)
-        }).join(' + ')
-        nodeValue = 'String(' + token + ')'
-    }
-    return '{\ntype: "#text",\nnodeType:3,\ndynamic:true,\nnodeValue: ' + nodeValue + '\n}'
-}
