+++ webpack.config.js
@@ -23,6 +23,7 @@ fix 在chrome与firefox下删掉select中的空白节点，会影响到selectedI
 ms-widget, ms-controller, ms-important生成的VM与对应的DOM都保存起来,
 并在avalon.vdomAdaptor中还原
 添加unescapeHTML与escapeHTML方法
+全新的lexer与 插值表达式抽取方法
      */
 })
 fs.writeFileSync('./src/seed/lang.share.js', text, 'utf8')
